XC-230: move timer logic to run_task
